Add test client and recording functionality
