+++ pywisetransfer/test/responses/currencies.yaml
@@ -1,0 +1,1 @@
+responses: []
